--- config/middleware/chatbot/request-context.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const configVars = {};
-
-configVars.containerProperty = 'body';
-configVars.requestContexts = {
-  keyword: 'User sends campaign keyword',
-  args: 'User send text input',
-  mms_image_url: 'User sends a picture',
-};
-module.exports = configVars;
