+++ config/middleware/chatbot/user-incoming-message.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const configVars = {};
+
+configVars.containerProperty = 'body';
+configVars.incomingMessageTypes = {
+  keyword: 'incoming keyword',
+  args: 'incoming input',
+  mms_image_url: 'incoming picture url',
+};
+module.exports = configVars;
