+++ lib/middleware/user-create.js
@@ -5,7 +5,7 @@ const helpers = require('../helpers');
 
 const User = require('../../app/models/User');
 
-module.exports = function registerNewUser() {
+module.exports = function createNewUser() {
   return (req, res, next) => {
     if (req.user) {
       return next();
